Changed naming to reflect project
