--- IISWithBindings/manifest.json
@@ -1,18 +1,0 @@
-﻿{
-  "extensions": {
-    "Keyfactor.Orchestrators.Extensions.IOrchestratorJobExtension": {
-      "CertStores.IISBindings.Inventory": {
-        "assemblypath": "IISWithBindings.dll",
-        "TypeFullName": "Keyfactor.Extensions.Orchestrator.IISWithBinding.Jobs.Inventory"
-      },
-      "CertStores.IISBindings.Management": {
-        "assemblypath": "IISWithBindings.dll",
-        "TypeFullName": "Keyfactor.Extensions.Orchestrator.IISWithBinding.Jobs.Management"
-      },
-      "CertStores.IISBindings.ReEnrollment": {
-        "assemblypath": "IISWithBindings.dll",
-        "TypeFullName": "Keyfactor.Extensions.Orchestrator.IISWithBinding.Jobs.ReEnrollment"
-      }
-    }
-  }
-}
